Put getStores in Application, don't pass it, move Command and CommandLineArguments from MainGazeTracker to Command, move some stuff around in GazeTrackerGtk
