+++ Application.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "OutputMethods.h"
+
 namespace Application {
 	// Tracker status
 	enum trackerStatus {
@@ -16,4 +18,7 @@ namespace Application {
 	extern int testDwelltimeParameter;
 	extern int sleepParameter;
 	extern CvRect *faceRectangle;
+
+	std::vector<boost::shared_ptr<AbstractStore> > getStores();
 }
+
